% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{avian_predation_pit_tag}
\alias{avian_predation_pit_tag}
\title{Avian Predation PIT-tag Recoveries}
\format{
A tibble with columns:
\describe{
  \item{location}{Waterbody where releases occurred (Upper Klamath Lake, Clear Lake Reservoir, etc.)}
  \item{fish_group}{Fish group (e.g., Adult LRS, Adult SNS, Juvenile suckers (SARP–Spr/Sum), Chinook, etc.)}
  \item{year}{Year of release / monitoring (2021–2023)}
  \item{available}{Number of PIT-tagged fish available to predators}
  \item{recovered}{Number of PIT-tags recovered on piscivorous waterbird colonies during the 2021–2023 breeding seasons}
}
}
\source{
Bird Research Northwest (2023)
[Avian Predation on UKB Suckers 2021–2023 Summary Report](https://www.birdresearchnw.org/Avian%20Predation%20on%20UKB%20Suckers_Summary%20Report%202021-2023.pdf)
}
\usage{
avian_predation_pit_tag
}
\description{
Data on the number of PIT-tagged fish that were available and subsequently
recovered on piscivorous waterbird colonies in the Upper Klamath Basin
during 2021–2023.
}
\details{
This dataset corresponds to **Table 2** in the report:
*Avian Predation on Upper Klamath Basin Suckers, 2021–2023 Summary Report*.
It provides raw counts of tagged fish released and tags recovered, by species
group, year, and location.
}
\keyword{datasets}
