% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estimate_predation_sarp}
\alias{estimate_predation_sarp}
\title{Predation Estimates on SARP and Chinook}
\format{

}
\source{
Bird Research Northwest (2023)
[Avian Predation on UKB Suckers 2021–2023 Summary Report](https://www.birdresearchnw.org/Avian%20Predation%20on%20UKB%20Suckers_Summary%20Report%202021-2023.pdf)
}
\usage{
estimate_predation_sarp
}
\description{
Estimates of predation rates (with 95% credible intervals) on PIT-tagged
Sucker Assisted Rearing Program (SARP) juvenile suckers and juvenile Chinook
Salmon released into the Upper Klamath Basin, Clear Lake Reservoir, and
Sheepy Lake.
}
\details{
This dataset corresponds to **Table 4** of the 2021–2023 summary report.
Estimates are provided separately for fish released in spring/summer (Spr/Sum)
and fall/winter (Fall/Win).
}
\keyword{datasets}
