% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fisheries_location_lookup}
\alias{fisheries_location_lookup}
\title{Fisheries Location Lookup}
\format{
A tibble with 9 rows and 10 columns
\itemize{
  \item \code{stream}: stream
  \item \code{sub_basin}: sub-basin name (upper klamath, lower klamath, trinity, shasta)
  \item \code{data_type}: type of data (rst, hatchery)
  \item \code{site_name}: name of site such as a RST site or hatchery location (big bar, shasta river, bogus, willow creek, pear creek, weitchpec, iron gate fish hatchery, trinity river hatchery, klamath hatchery)
  \item \code{agency}: agency that manages/monitors the site (arcata fwo, arcata fwo, cdfw, hoopa tribal fisheries department, karuk, yurok tribal fisheries program)
  \item \code{latitude}: longitude
  \item \code{longitude}: longitude
  \item \code{link}: web link containing more information about rotary screw trap
  }
}
\usage{
fisheries_location_lookup
}
\description{
This dataset compiles location data relevant to fisheries data collection efforts from across the Klamath Basin.
}
\keyword{datasets}
