% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{habitat_data}
\alias{habitat_data}
\title{Salmon Habitat Data}
\format{
A tibble with 15 rows and 13 columns
\itemize{
  \item \code{stream}: stream
  \item \code{sub_basin}: sub-basin name (Upper Klamath, Lower Klamath, Trinity)
  \item \code{location_name}: location name
  \item \code{model_type}: type of model used (2D hydrodynamic model, SRH-2D, micro-habitat models, HEC-EFM, SRH-1D)
  \item \code{length_miles}: geographical length of data coverage in miles
  \item \code{rm_start}: river mile of the beginning of extent, if applicable
  \item \code{rm_end}: river mile of the end of extent, if applicable
  \item \code{status}: status of model used (developed, in development)
  \item \code{location_souorce}: source used to determine the coordinates
  \item \code{source}: literature reference
  \item \code{link}: web link containing more information about data
  \item \code{latitude}: latitude of data location
  \item \code{longitude}: longitude of data location
  }
}
\usage{
habitat_data
}
\description{
This dataset compiles modeled salmon habitat data derived from literature reviews.
It includes information on habitat model extent, data sources, and approximate geographic locations.
}
\keyword{datasets}
