% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hatcheries}
\alias{hatcheries}
\title{Hatchery Locations}
\format{
A tibble with 3 rows and 8 columns
\itemize{
  \item \code{stream}: stream
  \item \code{sub_basin}: sub-basin name (Upper Klamath, Shasta)
  \item \code{data_type}: type of data (hatchery data)
  \item \code{site_name}: name of hatchery
  \item \code{agency}: agency that manages/monitors hatchery
  \item \code{latitude}: latitude
  \item \code{longitude}: longitude
  \item \code{link}: web link containing more information about hatchery
  }
}
\usage{
hatcheries
}
\description{
This dataset contains the hatchery locations and resources to find more information about them.
}
\keyword{datasets}
