% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historical_collections_releases}
\alias{historical_collections_releases}
\title{Historical Hatchery Collection Releases}
\format{
A tibble with 9 rows and 12 columns
\itemize{
  \item \code{hathcery_name}:
  \item \code{fiscal_year}: fiscal year
  \item \code{larvae_collected}: number of wild larval suckers collected
  \item \code{sarp_release}: number of fish released or transferred that were raised in the Sucker Assisted Rearing Program (SARP)
  \item \code{sarp_tl_mm}: average total length (TL) of the SARP release fish, measured in millimeters (mm)
  \item \code{fingerling_release}: number of fingerlings released or transferred
  \item \code{fingerling_tl_mm}: average total length (TL) of the fingerling release fish, measured in millimeters (mm)
  \item \code{fry_release}: number of fry released or transferred. This includes fry produced from experimental, wild-spawning activities that were transferred to other facilities (e.g., the Klamath Tribes Hatchery in FY24)
  \item \code{fry_tl_mm }: average total length (TL) of the  fry release fish, measured in millimeters (mm)
  \item \code{salvage_release}:  number of suckers salvaged throughout the year
  \item \code{salvage_sl_mm}: average length of release fish, measured in millimeters (mm)
  \item \code{salvage_tl_mm}: standard lengths (SL) of salvage fish are averages from measurements for each year
  }
}
\usage{
historical_collections_releases
}
\description{
Historical Hatchery Collection Releases
}
\keyword{datasets}
