% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KFNFH_annual_summary}
\alias{KFNFH_annual_summary}
\title{Annual Summary of Hatchery}
\format{
A tibble with 9 rows and 12 columns
\itemize{
  \item \code{hatchery_name}: name of hatchery (KFNFH)
  \item \code{fiscal_year}: fiscal year
  \item \code{larvae_collected}: number of wild larval suckers collected
  \item \code{sarp_release}: number of fish released or transferred that were raised in the Sucker Assisted Rearing Program (SARP)
  \item \code{sarp_tl_mm}: average total length (TL) of the SARP release fish, measured in millimeters (mm)
  \item \code{fingerling_release}: number of fingerlings released or transferred
  \item \code{fingerling_tl_mm}: average total length (TL) of the fingerling release fish, measured in millimeters (mm)
  \item \code{fry_release}: number of fry released or transferred. This includes fry produced from experimental, wild-spawning activities that were transferred to other facilities (e.g., the Klamath Tribes Hatchery in FY24)
  \item \code{fry_tl_mm }: average total length (TL) of the  fry release fish, measured in millimeters (mm)
  \item \code{salvage_release}:  number of suckers salvaged throughout the year
  \item \code{salvage_sl_mm}: average length of release fish, measured in millimeters (mm)
  \item \code{salvage_tl_mm}: standard lengths (SL) of salvage fish are averages from measurements for each year
  \item \code{primary_collection_method}: drift net, dip net, or both
  }
}
\source{
Klamath Falls National Fish Hatchery Annual Report for Fiscal Year 2024
}
\usage{
KFNFH_annual_summary
}
\description{
Summary of collections and releases from the Klamath Falls National Fish Hatchery since its inception in 2016. Data summarizes the annual inputs and outputs of the hatchery program by federal fiscal year
Corresponds to **Table 1** and **Table 4** in the FY2024 KFNFH Annual Report.
}
\keyword{datasets}
