% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KFNFH_hatchery_release}
\alias{KFNFH_hatchery_release}
\title{Hatchery Distributions and Transfers, KFNFH}
\format{
A tibble with 181 rows and 16 columns
\itemize{
  \item \code{hatchery_name}: name of hatchery (KFNFH)
  \item \code{fiscal_year}: fiscal year (2021-2025)
  \item \code{dispo_date}: date the fish distribution or transfer event occurred
  \item \code{species}: species of sucker being distributed (e.g. LRS, SNS, ESS LRS)
  \item \code{lot}: grouping or cohort of fish categorized primarily by their collection year (CY) or their operational origin
  \item \code{number_fish}: number of fish in that specific grouping (lot) that were distributed or transferred
  \item \code{weight_lb}: total weight of the fish lot being distributed, measured in pounds (lb)
  \item \code{actual_tl_mm}: actual total length (TL) of the fish, measured in millimeters (mm)
  \item \code{number_per_lb}: number of fish per pound for the specific lot being distributed
  \item \code{projected_tl_inches}: projected total length (TL) of the fish, measured in inches
  \item \code{projected_tl_mm}: projected total length (TL) of the fish, measured in millimeters (mm)
  \item \code{from_ponds}: source location within the hatchery facility from which the fish were harvested just prior to distribution or transfer
  \item \code{dispo_location}: ultimate destination where the fish were distributed or transferred
  \item \code{fish_type}: fish type as reported in the earlier (FY2021-2022) report tables
  \item \code{tl_inches}: total length in inches as reported in the earlier (FY2021-2022) report tables
  \item \code{tl_mm}: total length in millimeters as reported in the earlier (FY2021-2022) report tables
  }
}
\source{
Klamath Falls National Fish Hatchery Annual Reports for Fiscal Years 2021-2025
}
\usage{
KFNFH_hatchery_release
}
\description{
Summary of all fish distributions for repatriation or transfer from the Klamath Falls National Fish Hatchery during fiscal years 2021-2025. A "distribution" is considered a stocking or repatriation event, while a "transfer" involves moving fish for temporary holding.
Corresponds to **Table 15** in the FY2025 KFNFH Annual Report, **Table 11** in the FY2024 report, **Table 10** in the FY2023 report, **Table 5** in the FY2022 report, and **Table 2** in the FY2021 report.
}
\keyword{datasets}
