% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{megatable}
\alias{megatable}
\title{Megatable - IN DEVELOPMNENT}
\format{
A tibble with 4077 rows and 7 columns
\itemize{
  \item \code{location}:
  \item \code{subsection}:
  \item \code{section}:
  \item \code{species}:
  \item \code{lifestage}:
  \item \code{year}:
  \item \code{value}:
  }
}
\usage{
megatable
}
\description{
Digital version of [CDFW's Megatable](https://wildlife.ca.gov/Conservation/Fishes/Chinook-Salmon/Anadromous-Assessment)
}
\keyword{datasets}
