% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predation_estimates_avian_pit_tag}
\alias{predation_estimates_avian_pit_tag}
\title{Avian Predation PIT-tag Recoveries}
\format{
A tibble with 39 rows and 9 columns
\itemize{
  \item{location}{Waterbody where releases occurred (upper klamath lake, clear lake reservoir, etc.)}
  \item{species}{Fish species (e.g., lost river sucker, shortnose and klamath largescale suckers, sucker juveniles, chinook salmon,
  sucker, klamath largescale sucker, shortnose sucker)}
  \item{life_stage}{Fish life stage (e.g. adult, juvenile)}
  \item{origin}{Origin of realeased fish (e.g. wild, hatchery)}
  \item{release_season}{Season when fish were realeased (e.g. spring_summer, fall_winter)}
  \item{sarp_program}{Wheather fish are part of the Sucker Assisted Rearing Program (SARP) (e.g. TRUE, FALSE)}
  \item{year}{Year of release / monitoring (2021–2023)}
  \item{available}{Number of PIT-tagged fish available to predators}
  \item{recovered}{Number of PIT-tags recovered on piscivorous waterbird colonies during the 2021–2023 breeding seasons}
}
}
\source{
Bird Research Northwest (2023)
[Avian Predation on UKB Suckers 2021–2023 Summary Report](https://www.birdresearchnw.org/Avian%20Predation%20on%20UKB%20Suckers_Summary%20Report%202021-2023.pdf)
}
\usage{
predation_estimates_avian_pit_tag
}
\description{
Data on the number of PIT-tagged fish that were available and subsequently
recovered on piscivorous waterbird colonies in the Upper Klamath Basin
during 2021–2023.
This dataset corresponds to **Table 2** in the report:
*Avian Predation on Upper Klamath Basin Suckers, 2021–2023 Summary Report*.
It provides raw counts of tagged fish released and tags recovered, by species
group, year, and location.
}
\keyword{datasets}
