% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predation_estimates_hatchery}
\alias{predation_estimates_hatchery}
\title{Predation Estimates on SARP and Chinook}
\format{
A tibble with 18 rows and 10 columns
\itemize{
  \item{location}: Waterbody (Upper Klamath Lake, Clear Lake Reservoir, Sheepy Lake)
  \item{species}: Fish species (e.g. sucker, chinook salmon)
  \item{life_stage}: Fish life stage (e.g. juvenile)
  \item{origin}: Origin of realeased fish (e.g. hatchery)
  \item{release_season}: Season when fish were realeased (e.g. spring_summer, fall_winter)
  \item{sarp_program}: Wheather fish are part of the Sucker Assisted Rearing Program (SARP) (e.g. TRUE, FALSE)
  \item{year}: Year (2021–2023)
  \item{estimate_pct}: Estimated predation rate (% of available fish consumed)
  \item{lower_ci_pct}: Lower 95% credible interval
  \item{upper_ci_pct}: Upper 95% credible interval
}
}
\source{
Bird Research Northwest (2023)
[Avian Predation on UKB Suckers 2021–2023 Summary Report](https://www.birdresearchnw.org/Avian%20Predation%20on%20UKB%20Suckers_Summary%20Report%202021-2023.pdf)
}
\usage{
predation_estimates_hatchery
}
\description{
Estimates of predation rates (with 95% credible intervals) on PIT-tagged
Sucker Assisted Rearing Program (SARP) juvenile suckers and juvenile Chinook
Salmon released into the Upper Klamath Basin, Clear Lake Reservoir, and
Sheepy Lake.
This dataset corresponds to **Table 4** of the 2021–2023 summary report.
Estimates are provided separately for fish released in spring/summer (Spr/Sum)
and fall/winter (Fall/Win).
}
\keyword{datasets}
