% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rst_sites}
\alias{rst_sites}
\title{Rotary Screw Traps Data}
\format{
A tibble with 6 rows and 8 columns
\itemize{
  \item \code{stream}: stream
  \item \code{sub_basin}: sub-basin name (Upper Klamath, Lower Klamath, Trinity, Shasta)
  \item \code{data_type}: type of data (rotary screw trap data)
  \item \code{rst_name}: name of rotary screw trap
  \item \code{agency}: agency that manages/monitors screw trap
  \item \code{latitude}: longitude
  \item \code{longitude}: longitude
  \item \code{link}: web link containing more information about rotary screw trap
  }
}
\usage{
rst_sites
}
\description{
This dataset compiles rotary screw traps across the Klamath Basin.
}
\keyword{datasets}
