% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salmon_harvest}
\alias{salmon_harvest}
\title{Salmon Harvest}
\format{
A tibble with 392 rows and 8 columns
\itemize{
  \item \code{location}: location of data collection or estimate. This may be a river, wetland or marsh, hatchery, etc (klamath river, trinity river)
  \item \code{year}: return year of the escapement estimate/count (1978-2025)
  \item \code{species}: species name (spring chinook salmon, fall chinook salmon)
  \item \code{origin}: fish origin category (unknown)
  \item \code{lifestage}: life stage reported (adult, grilse)
  \item \code{estimate_type}: type of estimate reported. Work in progress as these have not been filled in yet for the data from the Megatables.
  \item \code{estimate}: value of estimate
  \item \code{source}: describes where data were sourced from. Currently data comes from: CDFW Spring and Fall Megatables, and Klamath River Technical Team Report 2026 obtained from Pacific Fishery Management Council
  }
}
\usage{
salmon_harvest
}
\description{
Digital version of \href{https://wildlife.ca.gov/Conservation/Fishes/Chinook-Salmon/Anadromous-Assessment}{CDFW's Spring Chinook Salmon Megatable},
\href{https://casalmon.org/wp-content/uploads/2024/08/FINAL-2023-Spring-Chinook-Megatable-v.28-Mar-2024.pdf}{CDFW's Spring Chinook Salmon Megatable} River Harvest sections,
and digitized tables (table 5 and appedix G.) from the
\href{https://www.pcouncil.org/documents/2026/06/2026-run-klamath-river-fall-chinook-salmon-age-specific-escapement-river-harvest-and-run-size-estimates-2026-run-may-20-2026.pdf/}{Klamath River Technical Team Report 2026} obtained from Pacific Fishery Management Council
}
\keyword{datasets}
