% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salmon_juvenile_abundance}
\alias{salmon_juvenile_abundance}
\title{Salmon Juvenile Abundance}
\format{
A tibble with 122 rows and 14 columns
\itemize{
  \item \code{julian_year}: julian year (2001-2022). Min/max years vary by location, species, type of estimate. Juvenile abundance typically spans multiple years (e.g. Oct 2020 - May 2021) and would be assigned a julian year for the latter part of this period.
  \item \code{stream}: stream ("scott river", "shasta river")
  \item \code{species}: species (coho salmon, steelhead, fall chinook salmon)
  \item \code{origin}: origin (wild)
  \item \code{lifestage}: lifestage category (yoy, age 1+, age 2+, smolt)
  \item \code{sex}: sex, if applicable to model (NA)
  \item \code{estimate_type}: type of model estimate (abundance, count)
  \item \code{estimate}: value of estimate
  \item \code{confidence_interval}: type of confidence or credible interval (e.g., 95)
  \item \code{lower_bounds_estimate}: value of lower bounds estimate, if applicable
  \item \code{upper_bounds_estimate}: value of upper bounds estimate, if applicable
  \item \code{estimation_method}: type of model used (mark-recapture)
  \item \code{is_complete_estimate}: data included in model are complete. In some cases, it is known that an
  estimate is not complete, otherwise it is assumed that estimate is complete.
  \item \code{source}: describes where data were sourced from. Currently there is one data source: (1) California Department of Fish and Wildlife (CDFW) document library (https://www.nrm.dfg.ca.gov/documents/ContextDocs.aspx?cat=Fisheries--AnadromousSalmonidPopulationMonitoring)
  }
}
\usage{
salmon_juvenile_abundance
}
\description{
This dataset compiles modeled estimates and credible intervals
from fisheries models such as abundance and survival estimates. This dataset
is not species or location specific. This is a key dataset!
In the future, this dataset may be broken down into separate tables like what
was done for spawner escapement.
}
\keyword{datasets}
