% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salmon_spawner_escapement}
\alias{salmon_spawner_escapement}
\title{Salmon Spawner Escapement}
\format{
A tibble with 1,854 rows and 13 columns
\itemize{
  \item \code{location}: River, hatchery, or basin where escapement is reported. ("bogus creek", "hoopa and yurok tributaries", "hoopa trinity tributaries", "iron gate hatchery", "klamath basin", "klamath river", "lower klamath river", "other klamath tributaries", "other klamath trinity tributaries", "other trinity tributaries", "salmon river", "scott river", "shasta river", "south fork trinity river", "trinity basin", "trinity river", "trinity river hatchery", "yurok klamath tributaries")
  \item \code{year}: Return year of the escapement estimate/count (1978-2025)
  \item \code{species}: Species name (spring chinook salmon, fall chinook salmon, coho salmon, winter steelhead, steelhead)
  \item \code{origin}: Fish origin category (hatchery, wild, mixed, unknown)
  \item \code{lifestage}: Life stage reported (adult, adult and subadult, grilse). In a few cases adult and subadult were not distinguished and are grouped together.
  \item \code{estimate_type}: Type of estimate reported (abundance, count, redd abundance). Work in progress as these have not been filled in yet for the data from the Megatables.
  \item \code{estimate}: value of estimate
  \item \code{confidence_interval}: type of confidence or credible interval (e.g., 95)
  \item \code{lower_bounds_estimate}: value of lower bounds estimate, if applicable
  \item \code{upper_bounds_estimate}: value of upper bounds estimate, if applicable
  \item \code{estimation_method}: type of model used (count, net upstream count, upstream count, expanded redd estimate, mark recapture. Work in progress as these have not been filled in yet for the data from the Megatables.)
  \item \code{is_complete_estimate}: data included in model are complete. In some cases, it is known that an
  estimate is not complete, otherwise it is assumed that estimate is complete.
  \item \code{source}: describes where data were sourced from. Currently there are 3 sources: (1) California Department of Fish and Wildlife (CDFW)
  \href{https://www.nrm.dfg.ca.gov/documents/ContextDocs.aspx?cat=Fisheries--AnadromousSalmonidPopulationMonitoring}{document library},
  (2) CDFW Spring and Fall Megatables, and (3) Klamath River Technical Team Report 2026 obtained from Pacific Fishery Management Council
  }
}
\usage{
salmon_spawner_escapement
}
\description{
Digital version of the Spawner Escapement sections from
\href{https://nrm.dfg.ca.gov/FileHandler.ashx?DocumentID=122850&inline}{CDFW's Fall Chinook Salmon Megatable}
and
\href{https://casalmon.org/wp-content/uploads/2024/08/FINAL-2023-Spring-Chinook-Megatable-v.28-Mar-2024.pdf}{CDFW's Spring Chinook Salmon Megatable},
as well as other data sources for Coho Salmon and Steelhead not included in the Megatables.

Coho Salmon, Steelhead, and Winter Steelhead data come from the California
Department of Fish and Wildlife (CDFW)
\href{https://www.nrm.dfg.ca.gov/documents/ContextDocs.aspx?cat=Fisheries--AnadromousSalmonidPopulationMonitoring}{document library}.

The most recent data come from digitized tables (Table 5 and Appendix G)
from the
\href{https://www.pcouncil.org/documents/2026/06/2026-run-klamath-river-fall-chinook-salmon-age-specific-escapement-river-harvest-and-run-size-estimates-2026-run-may-20-2026.pdf/}{Klamath River Technical Team Report 2026},
obtained from the Pacific Fishery Management Council.
}
\keyword{datasets}
