% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sucker_pit_tag_detection_2024}
\alias{sucker_pit_tag_detection_2024}
\title{Sucker PIT-tag Detections, FY2024}
\format{
A tibble with monitoring deployment summaries.
\itemize{
  \item \code{monitor_location}: PIT tag monitoring site
  \item \code{date_deployed}: deployment date
  \item \code{date_retrieved}: retrieval date
  \item \code{tags_total_unique}: total unique PIT tags detected
  \item \code{tags_sarp}: tags from SARP-reared fish
  \item \code{tags_salvage}: tags from salvaged fish
  \item \code{tags_unknown}: tags of unknown origin
}
}
\source{
Klamath Falls National Fish Hatchery Annual Report for Fiscal Year 2024
}
\usage{
sucker_pit_tag_detection_2024
}
\description{
Summary of PIT-tag detections from stationary monitoring locations
used to assess post-release survival and persistence of hatchery
and salvaged suckers in Upper Klamath Lake during fiscal year 2024.

Corresponds to **Table 12** in the FY2024 KFNFH Annual Report.
}
\keyword{datasets}
