% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{avian_predation_estimates}
\alias{avian_predation_estimates}
\title{Predation Estimates on SARP and Chinook}
\format{
A tibble with 39 rows and 16 columns
\itemize{
  \item{location}: Waterbody (upper klamath lake, clear lake reservoir, sheepy lake)
  \item{species}: Fish species (e.g. sucker, chinook salmon)
  \item{life_stage}: Fish life stage (e.g. juvenile)
  \item{origin}: Origin of realeased fish (e.g. hatchery, wild)
  \item{release_season}: Season when fish were realeased (e.g. spring_summer, fall_winter)
  \item{sarp_program}: Whether fish are part of the Sucker Assisted Rearing Program (SARP) (e.g. TRUE, FALSE)
  \item{year}: Year (2021–2023)
  \item{percent_estimate}: Estimated predation rate (% of available fish consumed)
  \item{lower_ci_pct}: Lower 95% credible interval
  \item{upper_ci_pct}: Upper 95% credible interval
  \item{number_adults_tagged}: Number of PIT-tagged fish available to predators
  \item{number_recovered_tags}: Number of PIT-tags recovered on piscivorous waterbird colonies during the 2021–2023 breeding seasons
}
}
\source{
Bird Research Northwest (2023)
[Avian Predation on UKB Suckers 2021–2023 Summary Report](https://birdresearchnw.org/brn-publications/avian-predation-on-ukb-suckers-summary-report-2021-2023/)
}
\usage{
avian_predation_estimates
}
\description{
This dataset combines avian predation rate estimates (with 95% credible intervals) and PIT-tag availability and recovery data for juvenile fishes in the Upper Klamath Basin. It includes PIT-tagged SARP suckers, hatchery-reared juvenile Chinook Salmon, and wild juvenile suckers released between 2021 and 2023.
This dataset corresponds to **Table 2**, **Table 3**, **Table 4**  of the 2021–2023 summary report.
}
\keyword{datasets}
