% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predation_estimates_wild}
\alias{predation_estimates_wild}
\title{Predation Estimates on Wild Suckers}
\format{

}
\source{
Bird Research Northwest (2023)
[Avian Predation on UKB Suckers 2021–2023 Summary Report](https://www.birdresearchnw.org/Avian%20Predation%20on%20UKB%20Suckers_Summary%20Report%202021-2023.pdf)
}
\usage{
predation_estimates_wild
}
\description{
Estimates of predation rates (with 95% credible intervals) on PIT-tagged
wild suckers (Lost River, Shortnose, Klamath Largescale, SNS–KLS hybrids,
and wild juveniles) by piscivorous colonial waterbirds in the Upper Klamath
Basin.
}
\details{
This dataset corresponds to **Table 3** of the 2021–2023 summary report.
Values are adjusted for PIT-tag detection and deposition probabilities.
}
\keyword{datasets}
